[CAKER-32] Part 1: Make scroll package compatible with macOS targets
- split integration view implementation into a macOS and nonMacOS ones
- implement macOS specific integration view (NSScrollView integration)
